requirements.txt
yt-dlp==2023.10.13
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
boto3==1.28.38
mutagen==1.46.0
Pillow==10.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.22